moved everything into tecaf.hpp
